fix @orgajs/theme-ui
