+++ packages/theme-ui/src/index.tsx
@@ -1,64 +1,44 @@
-import React from 'react'
-import { jsx } from 'theme-ui'
-import {
-  ThemeProvider as CoreProvider,
-  ThemeProviderProps as CoreThemeProviderProps,
+import { jsx } from '@theme-ui/core'
+export {
+  jsx,
+  __ThemeUIContext,
+  merge,
   useThemeUI,
-  __themeUiDefaultContextValue,
+  createElement,
 } from '@theme-ui/core'
-import { css, Theme } from '@theme-ui/css'
-import { OrgaProvider, OrgaProviderComponents } from './orga-provider'
-import { ColorModeProvider } from '@theme-ui/color-modes'
-import { Global } from '@emotion/react'
+export type {
+  ThemeUIContextValue,
+  IntrinsicSxElements,
+  SxProp,
+  ColorMode,
+  ColorModesScale,
+  CSSObject,
+  CSSProperties,
+  CSSPseudoSelectorProps,
+  ResponsiveStyleValue,
+  ThemeUICSSProperties,
+  ThemeUIStyleObject,
+  ThemeUICSSObject,
+  Theme,
+  ThemeStyles,
+  TLengthStyledSystem,
+  StylePropertyValue,
+} from '@theme-ui/core'
+export { useColorMode, InitializeColorMode } from '@theme-ui/color-modes'
+export * from '@theme-ui/components'
+export { css, get } from '@theme-ui/css'
 
-interface ThemeProviderProps extends Pick<CoreThemeProviderProps, 'theme'> {
-  children?: React.ReactNode
-  components?: OrgaProviderComponents
-}
-
-export { Themed, components } from './orga-provider'
-
-const RootStyles = () =>
-  jsx(Global, {
-    styles: (emotionTheme) => {
-      const theme = emotionTheme as Theme
-      const { useRootStyles } = theme.config || theme
-
-      if (useRootStyles === false || (theme.styles && !theme.styles.root)) {
-        return null
-      }
-
-      const boxSizing =
-        theme.config?.useBorderBox === false ? undefined : 'border-box'
-
-      return css({
-        '*': {
-          boxSizing,
-        },
-        html: {
-          variant: 'styles.root',
-        },
-        body: {
-          margin: 0,
-        },
-      })(theme)
+export const BaseStyles = (props: Record<string, unknown>) =>
+  jsx('div', {
+    ...props,
+    sx: {
+      fontFamily: 'body',
+      lineHeight: 'body',
+      fontWeight: 'body',
+      variant: 'styles',
     },
   })
 
-export const ThemeProvider: React.FC<ThemeProviderProps> = ({
-  theme,
-  components,
-  children,
-}) => {
-  const outer = useThemeUI()
-
-  const isTopLevel = outer === __themeUiDefaultContextValue
-  return (
-    <CoreProvider theme={theme}>
-      <ColorModeProvider>
-        {isTopLevel && <RootStyles />}
-        <OrgaProvider components={components}>{children}</OrgaProvider>
-      </ColorModeProvider>
-    </CoreProvider>
-  )
-}
+// orga version
+export { components, Themed } from './orga'
+export { ThemeProvider } from './theme-provider'
