+++ packages/theme-ui/src/orga.ts
@@ -10,7 +10,6 @@ import {
   FC,
   HTMLAttributes,
   ReactNode,
-  useEffect,
 } from 'react'
 
 type OrgaProviderComponentsKnownKeys = {
